fix source error

+++ nts.uk/uk.at/at.ctx/schedule/nts.uk.ctx.at.schedule.dom/src/main/java/nts/uk/ctx/at/schedule/dom/budget/external/actualresult/ExtBudgetTime.java
@@ -22,7 +22,7 @@ public class ExtBudgetTime extends TimeClockPrimitiveValue<ExtBudgetTime> {
      *
      * @param rawValue the raw value
      */
-    public ExtBudgetTime(Long rawValue) {
+    public ExtBudgetTime(int rawValue) {
         super(rawValue);
     }
